pyproject.toml
[tool.poetry]
name = "repl-nix-poke-star-discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_poke_star_discord_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
